+++ sdk/src/main/java/SourceClassMap.java
@@ -556,15 +556,12 @@ public class SourceClassMap {
 			put("org.zstack.sns.platform.snmp.SNSSnmpPlatformInventory", "org.zstack.sdk.sns.platform.snmp.SNSSnmpPlatformInventory");
 			put("org.zstack.sns.platform.wecom.SNSWeComAtPersonInventory", "org.zstack.sdk.sns.platform.wecom.SNSWeComAtPersonInventory");
 			put("org.zstack.sns.platform.wecom.SNSWeComEndpointInventory", "org.zstack.sdk.sns.platform.wecom.SNSWeComEndpointInventory");
+			put("org.zstack.sso.entity.CasAccountClientInventory", "org.zstack.sdk.CasAccountClientInventory");
 			put("org.zstack.sso.entity.OAuth2AccountClientInventory", "org.zstack.sdk.OAuth2AccountClientInventory");
 			put("org.zstack.sso.entity.OAuth2TokenInventory", "org.zstack.sdk.OAuth2TokenInventory");
 			put("org.zstack.sso.entity.SSOTokenInventory", "org.zstack.sdk.SSOTokenInventory");
 			put("org.zstack.sso.entity.SSOUrlTemplateInventory", "org.zstack.sdk.SSOUrlTemplateInventory");
-			put("org.zstack.sso.header.CasClientInventory", "org.zstack.sdk.CasClientInventory");
-			put("org.zstack.sso.header.CasState", "org.zstack.sdk.CasState");
-			put("org.zstack.sso.header.OAuth2ClientInventory", "org.zstack.sdk.OAuth2ClientInventory");
 			put("org.zstack.sso.header.RedirectUrlTemplate", "org.zstack.sdk.RedirectUrlTemplate");
-			put("org.zstack.sso.header.SSOClientInventory", "org.zstack.sdk.SSOClientInventory");
 			put("org.zstack.storage.backup.imagestore.ImageStoreBackupStorageInventory", "org.zstack.sdk.ImageStoreBackupStorageInventory");
 			put("org.zstack.storage.backup.imagestore.ImageStoreGcResult", "org.zstack.sdk.ImageStoreGcResult");
 			put("org.zstack.storage.backup.imagestore.ImageStoreImageStruct", "org.zstack.sdk.ImageStoreImageStruct");
@@ -820,8 +817,7 @@ public class SourceClassMap {
 			put("org.zstack.sdk.CCSCertificateInventory", "org.zstack.crypto.ccs.CCSCertificateInventory");
 			put("org.zstack.sdk.CCSCertificateUserRefInventory", "org.zstack.crypto.ccs.CCSCertificateUserRefInventory");
 			put("org.zstack.sdk.CCSCertificateUserState", "org.zstack.crypto.ccs.CCSCertificateUserState");
-			put("org.zstack.sdk.CasClientInventory", "org.zstack.sso.header.CasClientInventory");
-			put("org.zstack.sdk.CasState", "org.zstack.sso.header.CasState");
+			put("org.zstack.sdk.CasAccountClientInventory", "org.zstack.sso.entity.CasAccountClientInventory");
 			put("org.zstack.sdk.CdpPolicyInventory", "org.zstack.header.storage.cdp.CdpPolicyInventory");
 			put("org.zstack.sdk.CdpPolicyState", "org.zstack.header.storage.cdp.CdpPolicyState");
 			put("org.zstack.sdk.CdpTaskInventory", "org.zstack.header.storage.cdp.CdpTaskInventory");
@@ -1042,7 +1038,6 @@ public class SourceClassMap {
 			put("org.zstack.sdk.NvmeServerInventory", "org.zstack.storage.device.nvme.NvmeServerInventory");
 			put("org.zstack.sdk.NvmeTargetInventory", "org.zstack.storage.device.nvme.NvmeTargetInventory");
 			put("org.zstack.sdk.OAuth2AccountClientInventory", "org.zstack.sso.entity.OAuth2AccountClientInventory");
-			put("org.zstack.sdk.OAuth2ClientInventory", "org.zstack.sso.header.OAuth2ClientInventory");
 			put("org.zstack.sdk.OAuth2TokenInventory", "org.zstack.sso.entity.OAuth2TokenInventory");
 			put("org.zstack.sdk.OssBucketInventory", "org.zstack.header.aliyun.oss.OssBucketInventory");
 			put("org.zstack.sdk.OssBucketProperty", "org.zstack.header.aliyun.oss.OssBucketProperty");
@@ -1144,7 +1139,6 @@ public class SourceClassMap {
 			put("org.zstack.sdk.RouterAreaInventory", "org.zstack.header.protocol.RouterAreaInventory");
 			put("org.zstack.sdk.RunningState", "org.zstack.storage.device.localRaid.RunningState");
 			put("org.zstack.sdk.RunningTaskInfo", "org.zstack.header.core.progress.RunningTaskInfo");
-			put("org.zstack.sdk.SSOClientInventory", "org.zstack.sso.header.SSOClientInventory");
 			put("org.zstack.sdk.SSOTokenInventory", "org.zstack.sso.entity.SSOTokenInventory");
 			put("org.zstack.sdk.SSOUrlTemplateInventory", "org.zstack.sso.entity.SSOUrlTemplateInventory");
 			put("org.zstack.sdk.SchedulerJobGroupInventory", "org.zstack.header.scheduler.SchedulerJobGroupInventory");
