+++ sdk/src/main/java/org/zstack/sdk/CasAccountClientInventory.java
@@ -1,0 +1,119 @@
+package org.zstack.sdk;
+
+
+
+public class CasAccountClientInventory  {
+
+    public java.lang.String uuid;
+    public void setUuid(java.lang.String uuid) {
+        this.uuid = uuid;
+    }
+    public java.lang.String getUuid() {
+        return this.uuid;
+    }
+
+    public java.lang.String name;
+    public void setName(java.lang.String name) {
+        this.name = name;
+    }
+    public java.lang.String getName() {
+        return this.name;
+    }
+
+    public java.lang.String description;
+    public void setDescription(java.lang.String description) {
+        this.description = description;
+    }
+    public java.lang.String getDescription() {
+        return this.description;
+    }
+
+    public java.lang.String type;
+    public void setType(java.lang.String type) {
+        this.type = type;
+    }
+    public java.lang.String getType() {
+        return this.type;
+    }
+
+    public java.lang.String createAccountStrategy;
+    public void setCreateAccountStrategy(java.lang.String createAccountStrategy) {
+        this.createAccountStrategy = createAccountStrategy;
+    }
+    public java.lang.String getCreateAccountStrategy() {
+        return this.createAccountStrategy;
+    }
+
+    public java.lang.String deleteAccountStrategy;
+    public void setDeleteAccountStrategy(java.lang.String deleteAccountStrategy) {
+        this.deleteAccountStrategy = deleteAccountStrategy;
+    }
+    public java.lang.String getDeleteAccountStrategy() {
+        return this.deleteAccountStrategy;
+    }
+
+    public java.sql.Timestamp createDate;
+    public void setCreateDate(java.sql.Timestamp createDate) {
+        this.createDate = createDate;
+    }
+    public java.sql.Timestamp getCreateDate() {
+        return this.createDate;
+    }
+
+    public java.sql.Timestamp lastOpDate;
+    public void setLastOpDate(java.sql.Timestamp lastOpDate) {
+        this.lastOpDate = lastOpDate;
+    }
+    public java.sql.Timestamp getLastOpDate() {
+        return this.lastOpDate;
+    }
+
+    public java.lang.String loginMNUrl;
+    public void setLoginMNUrl(java.lang.String loginMNUrl) {
+        this.loginMNUrl = loginMNUrl;
+    }
+    public java.lang.String getLoginMNUrl() {
+        return this.loginMNUrl;
+    }
+
+    public java.lang.String casServerLoginUrl;
+    public void setCasServerLoginUrl(java.lang.String casServerLoginUrl) {
+        this.casServerLoginUrl = casServerLoginUrl;
+    }
+    public java.lang.String getCasServerLoginUrl() {
+        return this.casServerLoginUrl;
+    }
+
+    public java.lang.String casServerUrlPrefix;
+    public void setCasServerUrlPrefix(java.lang.String casServerUrlPrefix) {
+        this.casServerUrlPrefix = casServerUrlPrefix;
+    }
+    public java.lang.String getCasServerUrlPrefix() {
+        return this.casServerUrlPrefix;
+    }
+
+    public java.lang.String serverName;
+    public void setServerName(java.lang.String serverName) {
+        this.serverName = serverName;
+    }
+    public java.lang.String getServerName() {
+        return this.serverName;
+    }
+
+    public java.lang.String state;
+    public void setState(java.lang.String state) {
+        this.state = state;
+    }
+    public java.lang.String getState() {
+        return this.state;
+    }
+
+    public java.lang.String usernameProperty;
+    public void setUsernameProperty(java.lang.String usernameProperty) {
+        this.usernameProperty = usernameProperty;
+    }
+    public java.lang.String getUsernameProperty() {
+        return this.usernameProperty;
+    }
+
+}
