--- sdk/src/main/java/org/zstack/sdk/SSOClientInventory.java
@@ -1,87 +1,0 @@
-package org.zstack.sdk;
-
-
-
-public class SSOClientInventory  {
-
-    public java.lang.String uuid;
-    public void setUuid(java.lang.String uuid) {
-        this.uuid = uuid;
-    }
-    public java.lang.String getUuid() {
-        return this.uuid;
-    }
-
-    public java.lang.String name;
-    public void setName(java.lang.String name) {
-        this.name = name;
-    }
-    public java.lang.String getName() {
-        return this.name;
-    }
-
-    public java.lang.String description;
-    public void setDescription(java.lang.String description) {
-        this.description = description;
-    }
-    public java.lang.String getDescription() {
-        return this.description;
-    }
-
-    public java.lang.String clientType;
-    public void setClientType(java.lang.String clientType) {
-        this.clientType = clientType;
-    }
-    public java.lang.String getClientType() {
-        return this.clientType;
-    }
-
-    public java.lang.String loginType;
-    public void setLoginType(java.lang.String loginType) {
-        this.loginType = loginType;
-    }
-    public java.lang.String getLoginType() {
-        return this.loginType;
-    }
-
-    public java.lang.String loginMNUrl;
-    public void setLoginMNUrl(java.lang.String loginMNUrl) {
-        this.loginMNUrl = loginMNUrl;
-    }
-    public java.lang.String getLoginMNUrl() {
-        return this.loginMNUrl;
-    }
-
-    public java.lang.String redirectUrl;
-    public void setRedirectUrl(java.lang.String redirectUrl) {
-        this.redirectUrl = redirectUrl;
-    }
-    public java.lang.String getRedirectUrl() {
-        return this.redirectUrl;
-    }
-
-    public java.sql.Timestamp createDate;
-    public void setCreateDate(java.sql.Timestamp createDate) {
-        this.createDate = createDate;
-    }
-    public java.sql.Timestamp getCreateDate() {
-        return this.createDate;
-    }
-
-    public java.sql.Timestamp lastOpDate;
-    public void setLastOpDate(java.sql.Timestamp lastOpDate) {
-        this.lastOpDate = lastOpDate;
-    }
-    public java.sql.Timestamp getLastOpDate() {
-        return this.lastOpDate;
-    }
-
-    public java.lang.String accountUuid;
-    public void setAccountUuid(java.lang.String accountUuid) {
-        this.accountUuid = accountUuid;
-    }
-    public java.lang.String getAccountUuid() {
-        return this.accountUuid;
-    }
-
-}
